Change location of playerSelection.toLowerCase()
The call to this method was previously in playRound.
This had the undesirable effect of allowing "mixed case"
input to be displayed as part of the end of round message.

Also, as a result of moving the case conversion to
the getPlaySelection() function, the conversion call
that existed in playRound is now redundant.